+++ tests/testthat/helper_utils.R
@@ -4,7 +4,7 @@ testFile <- function(f, ..., text = FALSE) file.path(getTestDataPath(), paste0(f
 getTestAnaInfo <- function(path = patRoonData::exampleDataPath()) generateAnalysisInfo(path,
                                                                                        groups = c(rep("solvent", 3), rep("standard", 3)),
                                                                                        refs = "solvent")
-getTestFGroups <- function(anaInfo = getTestAnaInfo()) groupFeatures(findFeatures(anaInfo, "openms", logPath = NULL), "openms")
+getTestFGroups <- function(anaInfo = getTestAnaInfo(), ...) groupFeatures(findFeatures(anaInfo, "openms", logPath = NULL, ...), "openms")
 getEmptyTestFGroups <- function() getTestFGroups()[, "none"]
 getEmptyPLists <- function() MSPeakLists(algorithm = "none")