+++ NEWS.md
@@ -1,6 +1,7 @@
 # patRoon 0.1.0.9000
 
 ## January 2019
+* `filter()` method for components.
 * DataAnalysis formula generation: fixed neutral formula calculation if `MSMode="msms"`, now needs `adduct` argument.
 * Neutral loss filter for compounds.
 * **IMPORTANT** Adduct specification is now simplified and more generic by usage of a new `adduct` class. This means that `generateCompounds()` and `generateFormulas()` now expect slightly differing arguments. Please see their manual pages.